+++ config/jinja2.py
@@ -11,6 +11,7 @@ from django.utils.html import format_html, json_script
 from jinja2 import Environment
 from jinja_roos_components import setup_components
 
+from wies.core.roles import is_staff_user
 from wies.core.services.version import get_app_version
 
 
@@ -129,6 +130,7 @@ def environment(**options):
             "get_toggle_sort_url": get_toggle_sort_url,
             "get_sort_state": get_sort_state,
             "get_messages": get_messages,
+            "is_staff_user": is_staff_user,
             "DEBUG": settings.DEBUG,
             "APP_VERSION": get_app_version(),
         }